+++ src/service/recomposeFileHandler.ts
@@ -6,7 +6,8 @@ import * as path from 'node:path';
 
 import { Logger } from '@salesforce/core';
 import { CUSTOM_LABELS_FILE } from '../helpers/constants.js';
-import { composeAndWriteFile } from '../service/composeAndWriteFile.js';
+import { buildRecomposedFile } from './buildRecomposedFiles.js';
+
 const processFilesInDirectory = async (dirPath: string, metaSuffix: string): Promise<string[]> => {
   const combinedXmlContents: string[] = [];
   const files = await fs.readdir(dirPath);
@@ -33,7 +34,7 @@ const processFilesInDirectory = async (dirPath: string, metaSuffix: string): Pro
   return combinedXmlContents;
 };
 
-export async function composeFileHandler(
+export async function recomposeFileHandler(
   metaAttributes: {
     metaSuffix: string;
     xmlElement: string;
@@ -49,7 +50,7 @@ export async function composeFileHandler(
     const combinedXmlContents: string[] = await processFilesInDirectory(metadataPath, metaSuffix);
     const filePath = path.join(metadataPath, CUSTOM_LABELS_FILE);
 
-    await composeAndWriteFile(combinedXmlContents, filePath, xmlElement, log);
+    await buildRecomposedFile(combinedXmlContents, filePath, xmlElement, log);
   } else if (metaSuffix === 'bot' || metaSuffix === 'botVersion') {
     const botDirectories = (await fs.readdir(metadataPath)).map((file) => path.join(metadataPath, file));
 
@@ -73,7 +74,7 @@ export async function composeFileHandler(
               path.basename(botDirectory),
               `${subdirectoryBasename}.${metaSuffix}-meta.xml`
             );
-            await composeAndWriteFile(combinedXmlContents, filePath, xmlElement, log);
+            await buildRecomposedFile(combinedXmlContents, filePath, xmlElement, log);
           }
         }
       }
@@ -88,7 +89,7 @@ export async function composeFileHandler(
         const subdirectoryBasename = path.basename(subdirectory);
         const filePath = path.join(metadataPath, `${subdirectoryBasename}.${metaSuffix}-meta.xml`);
 
-        await composeAndWriteFile(combinedXmlContents, filePath, xmlElement, log);
+        await buildRecomposedFile(combinedXmlContents, filePath, xmlElement, log);
       }
     }
   }