+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# [2.0.0-beta.1](https://github.com/mcarvin8/sfdx-decomposer-plugin/compare/v1.1.2-beta.2...v2.0.0-beta.1) (2024-02-11)
+
+### Features
+
+- rename compose command ([6a7002e](https://github.com/mcarvin8/sfdx-decomposer-plugin/commit/6a7002e44f0b5c4d5cfd137e80673c3423f2fb2d))
+
+### BREAKING CHANGES
+
+- rename 'compose' command to 'recompose'
+
+## [1.1.2-beta.2](https://github.com/mcarvin8/sfdx-decomposer-plugin/compare/v1.1.2-beta.1...v1.1.2-beta.2) (2024-02-11)
+
+### Bug Fixes
+
+- **deps:** bump @salesforce/core from 6.5.1 to 6.5.3 ([0eff76c](https://github.com/mcarvin8/sfdx-decomposer-plugin/commit/0eff76c57039b697d9eaf9a482b24f90b8485399))
+
+## [1.1.2-beta.1](https://github.com/mcarvin8/sfdx-decomposer-plugin/compare/v1.1.1...v1.1.2-beta.1) (2024-02-11)
+
+### Bug Fixes
+
+- **deps:** bump fast-xml-parser from 4.3.3 to 4.3.4 ([654d939](https://github.com/mcarvin8/sfdx-decomposer-plugin/commit/654d9397d6d1c49557c673151898f0a380676c5c))
+
 ## [1.1.1](https://github.com/mcarvin8/sfdx-decomposer-plugin/compare/v1.1.0...v1.1.1) (2024-02-06)
 
 ### Bug Fixes